\name{Celsius to Fahrenheit}
\alias{Celsius to Fahrenheit}
\title{
Converts temperature units}
\description{
Converts an input temperature from Celsius to Fahrenheit}
\usage{
C_to_F(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Fahrenheit
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{C_to_F (32)
C_to_F (0)
  }
}
