\name{Celsius to Kelvin}
\alias{Celsius to Kelvin}
\title{
Converts temperature units}
\description{
Converts an input temperature from Celsius to Kelvin}
\usage{
C_to_K(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Kelvin
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{C_to_K (32)
C_to_K (0)
  }
}
