\name{Fahrenheit to Celsius}
\alias{Fahrenheit to Celsius}
\title{
Converts temperature units}
\description{
Converts an input temperature from Fahrenheit to Celsius}
\usage{
F_to_C(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Celsius
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{F_to_C (32)
F_to_C (0)
  }
}
