\name{Fahrenheit to Kelvin}
\alias{Fahrenheit to Kelvin}
\title{
Converts temperature units}
\description{
Converts an input temperature from Fahrenheit to Kelvin}
\usage{
F_to_K(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Kelvin
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{F_to_K (32)
F_to_K (0)
  }
}
