\name{Kelvin to Celsius}
\alias{Kelvin to Celsius}
\title{
Converts temperature units}
\description{
Converts an input temperature from Kelvin to Celsius}
\usage{
K_to_C(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Celsius
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{K_to_C (32)
K_to_C (0)
  }
}
