\name{Kelvin to Fahrenheit}
\alias{Kelvin to Fahrenheit}
\title{
Converts temperature units}
\description{
Converts an input temperature from Kelvin to Fahrenheit}
\usage{
K_to_F(x)
}
\arguments{
  \item{x}{a number}
}
\details{
}
\value{
Value returned is a number with units in Fahrenheit
}
\references{
}
\author{
Zach Jones}
\note{
}
\seealso{
}
\examples{
{K_to_F (32)
K_to_F (0)
  }
}
